clean up clean up everybody everywhere.
Bug fixes for the /user/appointments route. now send in the id of both the employee and customer no matter where its being hit from. only return appointments for the customer if the personal flag is true, meaning it's for a bookd customer.

Updates relating to that on all modals and pages